+++ platforms/nixos/modules/nmasur/presets/services/mealie/mealie.nix
@@ -10,14 +10,35 @@ in
   options.nmasur.presets.services.mealie.enable = lib.mkEnableOption "mealie recipe manager";
   config = lib.mkIf cfg.enable {
 
+    secrets.mealie-oidc-secret = {
+      source = ./mealie-oidc-secret.age;
+      dest = "${config.secretsDirectory}/mealie-oidc-secret";
+      prefix = "OIDC_CLIENT_SECRET=";
+    };
+    systemd.services.mealie-oidc-secret-secret = {
+      requiredBy = [ "mealie.service" ];
+      before = [ "mealie.service" ];
+    };
+
     services.mealie = {
       enable = true;
       port = 9099;
       database.createLocally = true;
       listenAddress = "127.0.0.1";
+      credentialsFile = config.secrets.mealie-oidc-secret.dest;
       settings = {
         TOKEN_TIME = 7200; # Hours for login to last (300 days)
+        OIDC_AUTH_ENABLED = "true";
+        OIDC_SIGNUP_ENABLED = "true";
+        OIDC_CONFIGURATION_URL = "https://${hostnames.auth}/.well-known/openid-configuration";
+        OIDC_CLIENT_ID = "040925ed-b39e-4442-b8e8-369c948c0cd2";
+        OIDC_PROVIDER_NAME = "Pocket ID";
+        OIDC_USER_CLAIM = "email";
       };
+    };
+
+    systemd.services.mealie = {
+      after = [ "mealie-oidc-secret-secret.service" ];
     };
 
     # Fix BASE_URL for downloading backups